+++ src/Control/Eff/Concurrent/Process/Interactive.hs
@@ -100,7 +100,7 @@ forkInteractiveScheduler ioScheduler = do
 killInteractiveScheduler :: SchedulerSession r -> IO ()
 killInteractiveScheduler (SchedulerSession qVar) =
   atomically (void (tryTakeTMVar qVar))
-
+                         
 -- | Send a 'Process' effect to the main process of a scheduler, this blocks
 -- until the effect is executed.
 submit